requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
contourpy==1.3.3
cycler==0.12.1
duckdb==1.5.0
fonttools==4.62.0
idna==3.11
kiwisolver==1.5.0
matplotlib==3.10.8
narwhals==2.18.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plotly==6.6.0
psycopg2-binary==2.9.11
py4j==0.10.9.9
pyparsing==3.3.2
pyspark==4.1.1
python-dateutil==2.9.0.post0
requests==2.32.5
seaborn==0.13.2
six==1.17.0
urllib3==2.6.3
